refactor(connector-tester): extract SnapshotIdMap from SnapshotComparer
